Changes made to work with the Data Engineering Day Pre-Lab Resources

+++ AnomalyDetection/RandomCutForest/src/main/java/software/amazon/flink/example/operator/InputDataMapper.java
@@ -21,4 +21,4 @@ package software.amazon.flink.example.operator;
 import org.apache.flink.api.common.functions.Function;
 
 @FunctionalInterface
-public interface InputDataMapper<T> extends Function, java.util.function.Function<T, float[]> {}
+public interface InputDataMapper<T> extends Function, java.util.function.Function<T, double[]> {}
